requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-quickbooks>=0.9.7
intuit-oauth>=1.2.4
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=4.9.0
pandas>=2.1.0
python-dateutil>=2.8.0
sqlalchemy>=2.0.0
alembic>=1.12.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4